+++ lib/app/features/battery_monitor/model/battery_monitor_model.dart
