+++ lib/app/features/battery_monitor/repository/battery_monitor_repository_fake.dart
@@ -1,4 +1,4 @@
-import 'package:penguin_kernel_manager/app/features/battery_monitor/model/battery_model.dart';
+import 'package:penguin_kernel_manager/app/features/battery_monitor/model/battery_monitor_model.dart';
 
 import 'package:penguin_kernel_manager/app/features/battery_monitor/repository/battery_monitor_repository.dart';
