+++ lib/app/features/gpu_monitor/model/gpu_monitor_model.dart
