+++ lib/app/features/gpu_monitor/repository/gpu_monitor_repository.dart
@@ -1,0 +1,5 @@
+import 'package:penguin_kernel_manager/app/features/gpu_monitor/model/gpu_monitor_model.dart';
+
+abstract class GpuRepository {
+  Future<GpuMonitorModel> getGpu();
+}
