+++ lib/app/features/gpu_monitor/repository/gpu_monitor_repository_impl.dart
@@ -1,11 +1,11 @@
 import 'dart:io';
 
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:penguin_kernel_manager/app/features/gpu_monitor/model/gpu_model.dart';
+import 'package:penguin_kernel_manager/app/features/gpu_monitor/model/gpu_monitor_model.dart';
 import 'package:penguin_kernel_manager/app/features/gpu_monitor/providers/gpu_providers.dart';
-import 'package:penguin_kernel_manager/app/features/gpu_monitor/providers/gpu_providers_android.dart';
-import 'package:penguin_kernel_manager/app/features/gpu_monitor/providers/gpu_providers_linux.dart';
-import 'package:penguin_kernel_manager/app/features/gpu_monitor/repository/gpu_repository.dart';
+import 'package:penguin_kernel_manager/app/features/gpu_monitor/providers/gpu_monitor_providers_android.dart';
+import 'package:penguin_kernel_manager/app/features/gpu_monitor/providers/gpu_monitor_providers_linux.dart';
+import 'package:penguin_kernel_manager/app/features/gpu_monitor/repository/gpu_monitor_repository.dart';
 
 class GpuRepositoryImpl extends GpuRepository {
   late final GpuMonitorProviders gpuProviders;
