--- lib/app/features/gpu_monitor/repository/gpu_repository.dart
@@ -1,5 +1,0 @@
-import 'package:penguin_kernel_manager/app/features/gpu_monitor/model/gpu_model.dart';
-
-abstract class GpuRepository {
-  Future<GpuMonitorModel> getGpu();
-}
