+++ lib/app/features/gpu_monitor/widget/gpu_monitor_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:penguin_kernel_manager/app/features/gpu_monitor/model/gpu_model.dart';
+import 'package:penguin_kernel_manager/app/features/gpu_monitor/model/gpu_monitor_model.dart';
 import 'package:penguin_kernel_manager/app/features/gpu_monitor/widget/gpu_frequency_widget.dart';
 import 'package:penguin_kernel_manager/app/features/gpu_monitor/widget/gpu_governor_widget.dart';
 import 'package:penguin_kernel_manager/app/features/gpu_monitor/widget/gpu_live_frequency_widget.dart';
